Move authentication.ts to /common/vscode/authentication.ts
